connectors code cleanup
